# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaymentCreate": {
        "additionalProperties": false,
        "properties": {
          "booking_id": {
            "format": "uuid",
            "title": "Booking Id",
            "type": "string"
          },
          "simulate_status": {
            "$ref": "#/components/schemas/PaymentStatus"
          }
        },
        "required": [
          "booking_id",
          "simulate_status"
        ],
        "title": "PaymentCreate",
        "type": "object"
      },
      "PaymentDetail": {
        "properties": {
          "amount_minor": {
            "title": "Amount Minor",
            "type": "integer"
          },
          "booking_id": {
            "format": "uuid",
            "title": "Booking Id",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/PaymentStatus"
          }
        },
        "required": [
          "id",
          "booking_id",
          "amount_minor",
          "currency",
          "status"
        ],
        "title": "PaymentDetail",
        "type": "object"
      },
      "PaymentStatus": {
        "enum": [
          "PENDING",
          "SUCCESS",
          "FAILED"
        ],
        "title": "PaymentStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WebhookRequest": {
        "additionalProperties": false,
        "example": {
          "amount_minor": 49900,
          "currency": "INR",
          "event_id": "evt-demo-payment-001",
          "payment_id": "11111111-1111-4111-8111-111111111111",
          "status": "SUCCESS"
        },
        "properties": {
          "amount_minor": {
            "examples": [
              49900
            ],
            "exclusiveMinimum": 0.0,
            "title": "Amount Minor",
            "type": "integer"
          },
          "currency": {
            "examples": [
              "INR"
            ],
            "pattern": "^INR$",
            "title": "Currency",
            "type": "string"
          },
          "event_id": {
            "examples": [
              "evt-demo-payment-001"
            ],
            "maxLength": 100,
            "minLength": 1,
            "title": "Event Id",
            "type": "string"
          },
          "payment_id": {
            "examples": [
              "11111111-1111-4111-8111-111111111111"
            ],
            "format": "uuid",
            "title": "Payment Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/PaymentStatus",
            "examples": [
              "SUCCESS"
            ]
          }
        },
        "required": [
          "event_id",
          "payment_id",
          "status",
          "amount_minor",
          "currency"
        ],
        "title": "WebhookRequest",
        "type": "object"
      },
      "WebhookResponse": {
        "properties": {
          "disposition": {
            "title": "Disposition",
            "type": "string"
          },
          "payment_id": {
            "format": "uuid",
            "title": "Payment Id",
            "type": "string"
          }
        },
        "required": [
          "payment_id",
          "disposition"
        ],
        "title": "WebhookResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      },
      "WebhookSecret": {
        "in": "header",
        "name": "X-Webhook-Secret",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Backend service for diagnostic test booking.",
    "title": "EVE Diagnostic Booking API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/payments/": {
      "post": {
        "operationId": "create_payments__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create",
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/webhook/": {
      "post": {
        "operationId": "webhook_payments_webhook__post",
        "parameters": [
          {
            "in": "header",
            "name": "x-webhook-secret",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Webhook-Secret"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "WebhookSecret": []
          }
        ],
        "summary": "Webhook",
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/{payment_id}/": {
      "get": {
        "operationId": "detail_payments__payment_id___get",
        "parameters": [
          {
            "in": "path",
            "name": "payment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Payment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Detail",
        "tags": [
          "payments"
        ]
      }
    },
...
  }
}